pyproject.toml
[tool.poetry]
name = "binal"
version = "0.1.0"
description = "A project to analyze Binance account balance."
authors = ["smd <andriipushkar@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiodns = "3.2.0"
aiofiles = "24.1.0"
aiogram = "3.19.0"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.10.11"
aiosignal = "1.3.2"
"annotated-types" = "0.7.0"
attrs = "25.3.0"
ccxt = "4.4.73"
certifi = "2025.1.31"
cffi = "1.17.1"
"charset-normalizer" = "3.4.1"
cryptography = "44.0.2"
dateparser = "1.2.1"
frozenlist = "1.5.0"
idna = "3.10"
"magic-filter" = "1.0.12"
multidict = "6.3.2"
numpy = "2.2.5"
pandas = "2.2.3"
propcache = "0.3.1"
pycares = "4.6.0"
pycparser = "2.22"
pycryptodome = "3.22.0"
pydantic = "2.10.6"
"pydantic_core" = "2.27.2"
"python-binance" = "1.0.28"
"python-dateutil" = "2.9.0.post0"
"python-dotenv" = "1.1.0"
pytz = "2025.2"
regex = "2024.11.6"
requests = "2.32.3"
setuptools = "78.1.0"
six = "1.17.0"
"typing-extensions" = "4.13.1"
tzdata = "2025.2"
tzlocal = "5.3.1"
urllib3 = "2.3.0"
websockets = "15.0.1"
yarl = "1.19.0"
pytest = "*"
"pytest-mock" = "*"
matplotlib = "*"
pandas-ta = "^0.3.14b0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
